# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Onboarding Diary API
  version: 1.0.0
  summary: Contract-first REST API for the Onboarding Diary application.
  description: |
    Normative API contract for the Onboarding Diary backend (Kotlin, Spring Boot 4.x,
    Spring WebFlux + coroutines, PostgreSQL).

    * Requirements: `docs/detailed-requirements.md` (each operation lists the
      `REQ-FUNC` ids it fulfils under `x-requirements`).
    * Authentication: email + password login returning a JWT bearer token
      (`bearerAuth`). Passwords are hashed exclusively in the backend (BCrypt / Argon2).
    * Every non-public operation returns `401 UNAUTHENTICATED` when the token is missing,
      invalid, expired, or belongs to a deactivated user. Clients clear their `AuthStore`
      and redirect to the login page on 401 (REQ-FUNC-006).
    * Every persisted resource exposes `createdAt` / `updatedAt` in UTC (REQ-FUNC-093).
    * The API is served from a Docker image selected by Spring profile (`dev`/`qa`/`prod`);
      CORS is permissive in `dev`/`qa` and allow-listed in `prod` (REQ-FUNC-095..098).
    * All errors share the `ErrorResponse` envelope. All list operations are paginated
      (`page`, `size`, `sort`) and return a `Page*` envelope.
    * Manager / Admin oversight uses the same entry endpoints with the `recruitId`
      query parameter. Managers may only pass recruits with an ACTIVE `Assignment` to them.
...
      currently assigned to that recruit only.
...
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users (admin)
      x-requirements: [REQ-FUNC-011, REQ-FUNC-091]
      parameters:
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/UserStatus' }
        - name: q
          in: query
          description: Case-insensitive match on email or fullName.
          schema: { type: string, maxLength: 100 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Users page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageUserSummary' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      operationId: createUser
      summary: Provision a user of any role without a password (admin); status INVITED
      x-requirements: [REQ-FUNC-012, REQ-FUNC-012a, REQ-FUNC-002, REQ-FUNC-003]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created with status INVITED; they complete setup via POST /auth/signup.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Email already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      operationId: getUser
      summary: Get user detail including current assignment (admin)
      x-requirements: [REQ-FUNC-013]
      responses:
        '200':
          description: User detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Update user profile fields, role and (while INVITED) email (admin)
      x-requirements: [REQ-FUNC-014, REQ-FUNC-012a]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUserUpdateRequest' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: Role change blocked by an ACTIVE assignment (`ROLE_CHANGE_BLOCKED_BY_ASSIGNMENT`), or email change on a non-INVITED user (`EMAIL_LOCKED`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{userId}/deactivate:
    post:
      tags: [Users]
      operationId: deactivateUser
      summary: Deactivate a user (admin); blocks login and invalidates tokens on next request
      x-requirements: [REQ-FUNC-015]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Deactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: Admin cannot deactivate own account (`CANNOT_DEACTIVATE_SELF`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{userId}/reactivate:
    post:
      tags: [Users]
      operationId: reactivateUser
      summary: Reactivate a deactivated user (admin)
      x-requirements: [REQ-FUNC-015]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Reactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token returned by `POST /auth/login` / `POST /auth/signup`. Sent as
        `Authorization: Bearer <token>`. Claims: sub (userId), email, role, iat, exp.
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index.
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      description: >-
        `field,asc|desc`; fields are whitelisted per resource. Defaults: entries
        (tasks, issues, feedback, notes) `entryDate,desc`; users `fullName,asc`;
        assignments `assignedAt,desc`; `/me/recruits` `fullName,asc`.
      schema: { type: string, pattern: '^[A-Za-z]+,(asc|desc)$' }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationFailed:
      description: Field-level validation failed or malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: Request validation failed
            details:
              - field: email
                code: INVALID_FORMAT
                message: must be a valid email address
            timestamp: '2026-09-27T06:00:00Z'
            path: /api/v1/auth/signup
    Unauthenticated:
      description: Missing, invalid or expired token, or deactivated user.
      headers:
        WWW-Authenticate: { schema: { type: string, const: Bearer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: UNAUTHENTICATED
            message: Authentication required
            details: []
            timestamp: '2026-09-27T06:00:00Z'
            path: /api/v1/me
    Forbidden:
      description: Role not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: FORBIDDEN
            message: You are not allowed to perform this action
            details: []
            timestamp: '2026-09-27T06:00:00Z'
            path: /api/v1/users
...
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: NOT_FOUND
            message: Resource not found
            details: []
            timestamp: '2026-09-27T06:00:00Z'
            path: /api/v1/tasks/1c2d
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Consistent error envelope for every non-2xx response (REQ-FUNC-090).
      required: [code, message, details, timestamp, path]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - MALFORMED_REQUEST
            - INVALID_CURRENT_PASSWORD
            - UNAUTHENTICATED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - NOT_ASSIGNED
            - NOT_FOUND
            - EMAIL_ALREADY_EXISTS
            - NOT_INVITED
            - ACCOUNT_ALREADY_ACTIVATED
            - EMAIL_LOCKED
            - ASSIGNMENT_UNCHANGED
            - CONFLICT
            - INVALID_STATE_TRANSITION
            - RESOLUTION_NOTES_REQUIRED
            - INVALID_ASSIGNMENT_PARTY
            - ROLE_CHANGE_BLOCKED_BY_ASSIGNMENT
            - CANNOT_DEACTIVATE_SELF
            - INTERNAL_ERROR
        message: { type: string }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        timestamp: { type: string, format: date-time }
        path: { type: string }
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        field: { type: string, description: JSON pointer-ish field name; absent for non-field errors. }
        code: { type: string, examples: [REQUIRED, INVALID_FORMAT, TOO_LONG, OUT_OF_RANGE, INVALID_ENUM, INVALID_TRANSITION] }
        message: { type: string }
    PageMeta:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }

    # ---------------- enums / value objects
    Role:
      type: string
      enum: [NEW_RECRUIT, MANAGER, ADMIN]
    UserStatus:
      type: string
      enum: [INVITED, ACTIVE, DEACTIVATED]
      description: INVITED = created by an Admin, no password yet, cannot log in.
    AssignmentStatus:
      type: string
      enum: [ACTIVE, REASSIGNED, ENDED]
      description: ENDED = closed because the recruit or manager was deactivated.
...
    Email:
      type: string
      format: email
      maxLength: 254
      description: Normalized (trimmed, lowercase) RFC-5322-style address; unique per user (REQ-FUNC-002/003).
      examples: [jane.doe@example.com]
...
    UserSummary:
      type: object
      required: [id, email, fullName, role, status]
      properties:
        id: { type: string, format: uuid }
        email: { $ref: '#/components/schemas/Email' }
        fullName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        status: { $ref: '#/components/schemas/UserStatus' }
        department: { type: [string, 'null'] }
        startDate: { type: [string, 'null'], format: date }
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          required: [createdAt, updatedAt, invitedAt]
          properties:
            invitedAt: { type: string, format: date-time }
            activatedAt: { type: [string, 'null'], format: date-time, description: Null while INVITED. }
            createdBy:
              description: Admin who provisioned the account; null for the bootstrap admin.
              oneOf:
                - $ref: '#/components/schemas/UserSummary'
                - type: 'null'
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          properties:
            currentAssignment:
              description: ACTIVE assignment when the user is a recruit; null otherwise.
              oneOf:
                - $ref: '#/components/schemas/Assignment'
                - type: 'null'
            activeRecruitCount:
              type: [integer, 'null']
              description: For managers, number of ACTIVE assignments.
...
    CreateUserRequest:
      type: object
      description: Creates an INVITED user without a password; the user sets it via POST /auth/signup.
      required: [email, fullName, role]
      additionalProperties: false
      properties:
        email: { $ref: '#/components/schemas/Email' }
        fullName: { type: string, minLength: 1, maxLength: 100 }
        role: { $ref: '#/components/schemas/Role' }
        department: { type: string, maxLength: 100 }
        startDate: { type: string, format: date }
    AdminUserUpdateRequest:
      type: object
      description: Admin-side update (REQ-FUNC-012a/014); absent properties are left unchanged, explicit null clears department/startDate.
      additionalProperties: false
      properties:
        fullName: { type: string, minLength: 1, maxLength: 100 }
        department: { type: [string, 'null'], maxLength: 100 }
        startDate: { type: [string, 'null'], format: date }
        role: { $ref: '#/components/schemas/Role' }
        email:
          allOf: [{ $ref: '#/components/schemas/Email' }]
          description: Only accepted while the user is INVITED (422 EMAIL_LOCKED otherwise).

    # ---------------- assignments
    Assignment:
      type: object
      required: [id, recruit, manager, assignedBy, status, assignedAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        recruit: { $ref: '#/components/schemas/UserSummary' }
        manager: { $ref: '#/components/schemas/UserSummary' }
        assignedBy: { $ref: '#/components/schemas/UserSummary' }
        status: { $ref: '#/components/schemas/AssignmentStatus' }
        assignedAt: { type: string, format: date-time }
        endedAt: { type: [string, 'null'], format: date-time }
        note: { type: [string, 'null'], maxLength: 500 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PageUserSummary:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/UserSummary' }
...